+++ src/index.ts
@@ -1,75 +1,67 @@
 import 'source-map-support/register';
 
-import * as http from 'http';
+import * as path from 'path';
+import * as child_process from 'child_process';
+import * as commander from 'commander';
 
-import { hb } from './hb';
-import { users } from './users';
+let homebridge;
 
-export = (homebridge) => {
-  hb.homebridge = homebridge;
+export = (api) => {
+  homebridge = api;
   homebridge.registerPlatform('homebridge-config-ui-x', 'config', HomebridgeConfigUi);
 };
 
 class HomebridgeConfigUi {
-  private server;
+  log;
 
   constructor(log, config) {
-    // setup
-    hb.init(log, config);
+    this.log = log;
 
-    // bootstrap the server
-    this.bootstrap();
-  }
+    const setup = {
+      homebridgeVersion: homebridge.serverVersion,
+      configPath: homebridge.user.configPath(),
+      storagePath: homebridge.user.storagePath(),
+      config: config
+    };
 
-  async bootstrap() {
-    // ensure auth.json is setup correctly
-    await users.setupAuthFile();
+    commander
+      .allowUnknownOption()
+      .option('-P, --plugin-path [path]', '', (p) => config.pluginPath = p)
+      .option('-I, --insecure', '', () => config.homebridgeInsecure = true)
+      .parse(process.argv);
 
-    // load config.json into memory
-    await hb.refreshHomebridgeConfig();
-
-    // dynamically load modules so app is only loaded if plugin is enabled
-    const { ExpressServer } = await import('./app');
-    const { WSS } = await import('./wss');
-
-    const app = new ExpressServer().app;
-
-    this.server = http.createServer(app);
-
-    // attach websocker server to the express server
-    hb.wss = new WSS(this.server);
-
-    this.server.listen(hb.port);
-    this.server.on('error', this.onServerError.bind(this));
-    this.server.on('listening', this.onServerListening.bind(this));
-  }
-
-  onServerListening() {
-    const addr = this.server.address();
-    const bind = typeof addr === 'string' ? 'pipe ' + addr : 'port ' + addr.port;
-    const msg = 'Console is listening on ' + bind + '.';
-    hb.log(msg);
-  }
-
-  onServerError(error) {
-    if (error.syscall !== 'listen') {
-      throw error;
+    if (config.noFork) {
+      this.noFork(setup);
+    } else {
+      this.fork(setup);
     }
+  }
 
-    const bind = typeof hb.port === 'string' ? 'Pipe ' + hb.port : 'Port ' + hb.port;
+  /**
+   * Run plugin as a seperate node.js process
+   */
+  fork(setup) {
+    const ui = child_process.fork(path.resolve(__dirname, 'bin/fork'));
 
-    switch (error.code) {
-      case 'EACCES':
-        console.error(bind + ' requires elevated privileges');
-        process.exit(1);
-        break;
-      case 'EADDRINUSE':
-        console.error(bind + ' is already in use');
-        process.exit(1);
-        break;
-      default:
-        throw error;
-    }
+    this.log(`Spawning homebridge-config-ui-x with PID`, ui.pid);
+
+    ui.on('message', (message) => {
+      if (message === 'ready') {
+        ui.send(setup);
+      }
+    });
+
+    ui.on('close', () => {
+      process.exit(1);
+    });
+  }
+
+  /**
+   * Run plugin in the main homebridge process
+   */
+  async noFork(setup) {
+    const { UiServer } = await import('./server');
+    return new UiServer(setup);
   }
 
   accessories(callback) {